refactor: extract backend and emulator modules from electron/main
Split electron/main.ts into focused modules:
- backend.ts: Python backend process management
- emulatorDetect.ts: Android emulator detection logic
Simplify main.ts entry point
